@@ -1,10 +1,10 @@
 {
   "name": "terriajs",
-  "version": "8.3.7",
+  "version": "8.4.0",
   "description": "Geospatial data visualization platform.",
   "license": "Apache-2.0",
   "engines": {
-    "node": ">= 14.0.0"
+    "node": ">= 16.0.0"
   },
   "repository": {
     "type": "git",
@@ -217,13 +217,13 @@
     "klaw-sync": "^4.0.0",
     "minimist": "^1.2.8",
     "node-notifier": "^5.1.2",
-    "node-sass": "^6.0.1",
     "plugin-error": "^1.0.1",
     "prettier": "2.7.1",
     "pretty-quick": "^1.10.0",
     "react-shallow-testutils": "^3.0.0",
     "react-test-renderer": "^16.3.2",
     "regenerator-runtime": "^0.13.2",
+    "sass": "^1.66.1",
     "terriajs-server": "^4.0.0",
     "yaml": "^1.10.0"
   },
