@@ -27,7 +27,7 @@ or with overrides on icon
 
 import React from "react";
 import styled, { useTheme } from "styled-components";
-const Box: any = require("./Box").default;
+import Box from "./Box";
 import { GLYPHS, StyledIcon } from "./Icon";
 
 const StyledSelect = styled.select<SelectProps>`
