@@ -1,4 +1,4 @@
-import { i18n } from "i18next";
+import { TOptions, i18n } from "i18next";
 import { isJsonString } from "../Core/Json";
 
 export const TRANSLATE_KEY_PREFIX = "translate#";
@@ -17,12 +17,12 @@ export const TRANSLATE_KEY_PREFIX = "translate#";
 export function applyTranslationIfExists(
   keyOrString: string,
   i18n: i18n,
-  options?: { [key: string]: string }
+  options?: TOptions
 ): string {
   // keyOrString could be undefined in some cases even if we type it as string
   if (isJsonString(keyOrString as unknown)) {
     if (keyOrString.indexOf(TRANSLATE_KEY_PREFIX) === 0) {
-      const translationKey = keyOrString.substr(TRANSLATE_KEY_PREFIX.length);
+      const translationKey = keyOrString.substring(TRANSLATE_KEY_PREFIX.length);
       return i18n.exists(translationKey)
         ? i18n.t(translationKey, options)
         : translationKey;
