@@ -3,13 +3,13 @@ import { computed, makeObservable } from "mobx";
 import CesiumMath from "terriajs-cesium/Source/Core/Math";
 import Rectangle from "terriajs-cesium/Source/Core/Rectangle";
 import URI from "urijs";
+import { JsonObject, isJsonArray, isJsonString } from "../../../Core/Json";
+import TerriaError from "../../../Core/TerriaError";
 import containsAny from "../../../Core/containsAny";
 import createDiscreteTimesFromIsoSegments from "../../../Core/createDiscreteTimes";
 import filterOutUndefined from "../../../Core/filterOutUndefined";
 import isDefined from "../../../Core/isDefined";
 import isReadOnlyArray from "../../../Core/isReadOnlyArray";
-import { isJsonArray, isJsonString, JsonObject } from "../../../Core/Json";
-import TerriaError from "../../../Core/TerriaError";
 import { terriaTheme } from "../../../ReactViews/StandardUserInterface/StandardTheme";
 import {
   InfoSectionTraits,
@@ -19,6 +19,7 @@ import {
   KeyValueTraits,
   WebCoverageServiceParameterTraits
 } from "../../../Traits/TraitsClasses/ExportWebCoverageServiceTraits";
+import { FeatureInfoTemplateTraits } from "../../../Traits/TraitsClasses/FeatureInfoTraits";
 import LegendTraits from "../../../Traits/TraitsClasses/LegendTraits";
 import { RectangleTraits } from "../../../Traits/TraitsClasses/MappableTraits";
 import WebMapServiceCatalogItemTraits, {
@@ -29,18 +30,18 @@ import WebMapServiceCatalogItemTraits, {
   WebMapServiceAvailableLayerStylesTraits,
   WebMapServiceAvailableStyleTraits
 } from "../../../Traits/TraitsClasses/WebMapServiceCatalogItemTraits";
-import createStratumInstance from "../../Definition/createStratumInstance";
 import LoadableStratum from "../../Definition/LoadableStratum";
 import Model, { BaseModel } from "../../Definition/Model";
 import StratumFromTraits from "../../Definition/StratumFromTraits";
+import createStratumInstance from "../../Definition/createStratumInstance";
 import proxyCatalogItemUrl from "../proxyCatalogItemUrl";
 import { CapabilitiesStyle } from "./OwsInterfaces";
 import WebMapServiceCapabilities, {
   CapabilitiesContactInformation,
   CapabilitiesDimension,
   CapabilitiesLayer,
-  getRectangleFromLayer,
-  MetadataURL
+  MetadataURL,
+  getRectangleFromLayer
 } from "./WebMapServiceCapabilities";
 import WebMapServiceCatalogItem from "./WebMapServiceCatalogItem";
 
@@ -720,13 +721,33 @@ export default class WebMapServiceCapabilitiesStratum extends LoadableStratum(
 
   @computed
   get supportsGetLegendGraphic(): boolean {
+    const capabilities = this.capabilities?.json?.Capability;
+
     return (
       isDefined(this.capabilities?.json?.["xmlns:sld"]) ||
-      isDefined(
-        this.capabilities?.json?.Capability?.Request?.GetLegendGraphic
-      ) ||
+      isDefined(capabilities?.Request?.GetLegendGraphic) ||
+      (Array.isArray(capabilities?.ExtendedCapabilities?.ExtendedRequest) &&
+        capabilities.ExtendedCapabilities.ExtendedRequest.find(
+          (r: JsonObject) => r?.Request === "GetLegendGraphic"
+        )) ||
       (this.catalogItem.isGeoServer ?? false) ||
       (this.catalogItem.isNcWMS ?? false)
+    );
+  }
+
+  @computed
+  get supportsGetTimeseries() {
+    // Don't use GetTimeseries if there is only one timeslice
+    if ((this.catalogItem.discreteTimes?.length ?? 0) <= 1) return false;
+
+    const capabilities = this.capabilities?.json?.Capability;
+
+    return !!(
+      isDefined(capabilities?.Request?.GetTimeseries) ||
+      (Array.isArray(capabilities?.ExtendedCapabilities?.ExtendedRequest) &&
+        capabilities.ExtendedCapabilities.ExtendedRequest.find(
+          (r: JsonObject) => r?.Request === "GetTimeseries"
+        ))
     );
   }
 
@@ -841,6 +862,8 @@ export default class WebMapServiceCapabilitiesStratum extends LoadableStratum(
    * - Plain text
    *
    * If no matching format can be found in GetCapabilities, then Cesium will use defaults (see `WebMapServiceImageryProvider.DefaultGetFeatureInfoFormats`)
+   *
+   * If supportsGetTimeseries, use CSV
    */
   @computed get getFeatureInfoFormat():
     | StratumFromTraits<GetFeatureInfoFormat>
@@ -853,6 +876,10 @@ export default class WebMapServiceCapabilitiesStratum extends LoadableStratum(
       : isJsonString(formats)
       ? [formats]
       : [];
+
+    if (this.catalogItem.supportsGetTimeseries) {
+      return { format: "text/csv", type: "text" };
+    }
 
     if (formatsArray.includes("application/json"))
       return { format: "application/json", type: "json" };
@@ -877,6 +904,26 @@ export default class WebMapServiceCapabilitiesStratum extends LoadableStratum(
     }
     if (formatsArray.includes("text/plain"))
       return { format: "text/plain", type: "text" };
+  }
+
+  /** If supportsGetTimeseries, override the "request" parameter in GetFeatureInfo to be "GetTimeseries".
+   * We also set time to empty, so we get values for all times (as opposed to just the current time)
+   */
+  @computed get getFeatureInfoParameters() {
+    if (this.catalogItem.supportsGetTimeseries) {
+      return { request: "GetTimeseries", time: "" };
+    }
+    return undefined;
+  }
+
+  /** If getFeatureInfoFormat is text/csv, set featureInfoTemplate to show chart. */
+  @computed
+  get featureInfoTemplate() {
+    if (this.catalogItem.getFeatureInfoFormat.format === "text/csv")
+      return createStratumInstance(FeatureInfoTemplateTraits, {
+        template: `{{terria.timeSeries.chart}}`,
+        showFeatureInfoDownloadWithTemplate: true
+      });
   }
 
   @computed get linkedWcsParameters() {
