@@ -8,18 +8,18 @@
 //  Solution: think in terms of pipelines with computed observables, document patterns.
 // 4. All code for all catalog item types needs to be loaded before we can do anything.
 import i18next from "i18next";
-import { computed, runInAction, makeObservable, override } from "mobx";
-import combine from "terriajs-cesium/Source/Core/combine";
+import { computed, makeObservable, override, runInAction } from "mobx";
 import GeographicTilingScheme from "terriajs-cesium/Source/Core/GeographicTilingScheme";
 import JulianDate from "terriajs-cesium/Source/Core/JulianDate";
 import WebMercatorTilingScheme from "terriajs-cesium/Source/Core/WebMercatorTilingScheme";
+import combine from "terriajs-cesium/Source/Core/combine";
 import GetFeatureInfoFormat from "terriajs-cesium/Source/Scene/GetFeatureInfoFormat";
 import WebMapServiceImageryProvider from "terriajs-cesium/Source/Scene/WebMapServiceImageryProvider";
 import URI from "urijs";
+import TerriaError from "../../../Core/TerriaError";
 import createTransformerAllowUndefined from "../../../Core/createTransformerAllowUndefined";
 import filterOutUndefined from "../../../Core/filterOutUndefined";
 import isDefined from "../../../Core/isDefined";
-import TerriaError from "../../../Core/TerriaError";
 import CatalogMemberMixin, {
   getName
 } from "../../../ModelMixins/CatalogMemberMixin";
@@ -32,6 +32,10 @@ import MappableMixin, {
 import MinMaxLevelMixin from "../../../ModelMixins/MinMaxLevelMixin";
 import TileErrorHandlerMixin from "../../../ModelMixins/TileErrorHandlerMixin";
 import UrlMixin from "../../../ModelMixins/UrlMixin";
+import {
+  TimeSeriesFeatureInfoContext,
+  csvFeatureInfoContext
+} from "../../../Table/tableFeatureInfoContext";
 import WebMapServiceCatalogItemTraits, {
   SUPPORTED_CRS_3857,
   SUPPORTED_CRS_4326
@@ -41,6 +45,7 @@ import CreateModel from "../../Definition/CreateModel";
 import LoadableStratum from "../../Definition/LoadableStratum";
 import { BaseModel } from "../../Definition/Model";
 import StratumOrder from "../../Definition/StratumOrder";
+import FeatureInfoContext from "../../Feature/FeatureInfoContext";
 import SelectableDimensions, {
   SelectableDimensionEnum
 } from "../../SelectableDimensions/SelectableDimensions";
@@ -49,6 +54,7 @@ import proxyCatalogItemUrl from "../proxyCatalogItemUrl";
 import WebMapServiceCapabilities from "./WebMapServiceCapabilities";
 import WebMapServiceCapabilitiesStratum from "./WebMapServiceCapabilitiesStratum";
 import WebMapServiceCatalogGroup from "./WebMapServiceCatalogGroup";
+import TerriaFeature from "../../Feature/Feature";
 
 /** This LoadableStratum is responsible for setting WMS version based on CatalogItem.url */
 export class WebMapServiceUrlStratum extends LoadableStratum(
@@ -94,7 +100,7 @@ class WebMapServiceCatalogItem
       )
     )
   )
-  implements SelectableDimensions
+  implements SelectableDimensions, FeatureInfoContext
 {
   /**
    * The collection of strings that indicate an Abstract property should be ignored.  If these strings occur anywhere
@@ -514,8 +520,9 @@ class WebMapServiceCatalogItem
           (this.maximumShownFeatureInfos ??
             this.terria.configParameters.defaultMaximumShownFeatureInfos),
         ...this.parameters,
-        ...this.getFeatureInfoParameters,
-        ...dimensionParameters
+        // Note order is important here, as getFeatureInfoParameters may override `time` dimension value
+        ...dimensionParameters,
+        ...this.getFeatureInfoParameters
       };
 
       const diffModeParameters = this.isShowingDiff
@@ -526,10 +533,9 @@ class WebMapServiceCatalogItem
         parameters.COLORSCALERANGE = this.colorScaleRange;
       }
 
-      if (isDefined(this.styles)) {
-        parameters.styles = this.styles;
-        getFeatureInfoParameters.styles = this.styles;
-      }
+      // Styles parameter is mandatory (for GetMap and GetFeatureInfo requests), but can be empty string to use default style
+      parameters.styles = this.styles ?? "";
+      getFeatureInfoParameters.styles = this.styles ?? "";
 
       Object.assign(parameters, diffModeParameters);
 
@@ -744,6 +750,15 @@ class WebMapServiceCatalogItem
       ...this.wmsDimensionSelectableDimensions,
       ...this.styleSelectableDimensions
     ]);
+  }
+
+  /** If GetFeatureInfo/GetTimeseries request is returning CSV, we need to parse it into TimeSeriesFeatureInfoContext.
+   */
+  @computed get featureInfoContext(): (
+    feature: TerriaFeature
+  ) => TimeSeriesFeatureInfoContext {
+    if (this.getFeatureInfoFormat.format !== "text/csv") return () => ({});
+    return csvFeatureInfoContext(this);
   }
 }
 
