@@ -1,17 +1,22 @@
-import { observable, makeObservable } from "mobx";
-import SearchResult from "./SearchResult";
+import { makeObservable, observable } from "mobx";
 import { IPromiseBasedObservable, fromPromise } from "mobx-utils";
-import SearchProvider from "./SearchProvider";
+import SearchProviderMixin from "../../ModelMixins/SearchProviders/SearchProviderMixin";
+import SearchResult from "./SearchResult";
 
 export default class SearchProviderResults {
   @observable results: SearchResult[] = [];
-  @observable message: string | undefined;
+  @observable message?: {
+    content: string;
+    params?: {
+      [key: string]: string | number | undefined;
+    };
+  };
   isCanceled = false;
   resultsCompletePromise: IPromiseBasedObservable<void> = fromPromise(
     Promise.resolve()
   );
 
-  constructor(readonly searchProvider: SearchProvider) {
+  constructor(readonly searchProvider: SearchProviderMixin.Instance) {
     makeObservable(this);
   }
 
