@@ -1,13 +1,7 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'footer': string;
-  'heading': string;
-  'isOpen': string;
-  'items': string;
-  'light': string;
-  'provider-result': string;
-  'providerResult': string;
+  'legends': string;
 }
 declare var cssExports: CssExports;
 export = cssExports;
