@@ -1,15 +1,11 @@
 import React from "react";
 import { observer } from "mobx-react";
-
 import PropTypes from "prop-types";
 import { withTranslation } from "react-i18next";
-
 import defined from "terriajs-cesium/Source/Core/defined";
-
-import DataCatalogMember from "./DataCatalogMember";
 import SearchHeader from "../Search/SearchHeader";
-
 import Styles from "./data-catalog.scss";
+import DataCatalogMember from "./DataCatalogMember";
 
 // Displays the data catalog.
 @observer
@@ -43,7 +39,7 @@ class DataCatalog extends React.Component {
           <>
             <label className={Styles.label}>{t("search.resultsLabel")}</label>
             <SearchHeader
-              searchResults={catalogSearchProvider}
+              searchResults={searchState.catalogSearchResults}
               isWaitingForSearchToStart={
                 searchState.isWaitingToStartCatalogSearch
               }
