@@ -1,10 +1,7 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'formLabel': string;
-  'searchClear': string;
-  'searchData': string;
-  'searchField': string;
+  'btn': string;
 }
 declare var cssExports: CssExports;
 export = cssExports;
