@@ -36,8 +36,7 @@ beforeEach(function () {
 
   viewState = new ViewState({
     terria: terria,
-    catalogSearchProvider: null,
-    locationSearchProviders: []
+    catalogSearchProvider: undefined
   });
 });
 
