@@ -15,8 +15,7 @@ describe("Branding", function () {
     terria = new Terria();
     viewState = new ViewState({
       terria,
-      catalogSearchProvider: undefined,
-      locationSearchProviders: []
+      catalogSearchProvider: undefined
     });
   });
 
