@@ -15,8 +15,7 @@ describe("ToolButton", function () {
     const terria = new Terria();
     viewState = new ViewState({
       terria,
-      catalogSearchProvider: undefined,
-      locationSearchProviders: []
+      catalogSearchProvider: undefined
     });
   });
 });
