@@ -16,8 +16,7 @@ describe("ViewingControls", function () {
     terria = new Terria();
     viewState = new ViewState({
       terria,
-      catalogSearchProvider: undefined,
-      locationSearchProviders: []
+      catalogSearchProvider: undefined
     });
   });
 
