@@ -12,8 +12,7 @@ describe("FeatureInfoPanel", function () {
     terria = new Terria();
     viewState = new ViewState({
       terria,
-      catalogSearchProvider: undefined,
-      locationSearchProviders: []
+      catalogSearchProvider: undefined
     });
   });
 
